refactor: useGetClass 분리, mutation만 관리하는 useClassMutation으로 변경
